backend_python/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.5.0
PyJWT==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
requests==2.31.0
